Move external link and translated-external-link
