+++ src/components/landing/layout/footer/powered-by-links.tsx
@@ -2,8 +2,8 @@ import React, { Fragment } from 'react'
 import { Trans, useTranslation } from 'react-i18next'
 import { useSelector } from 'react-redux'
 import { ApplicationState } from '../../../../redux'
-import { ExternalLink } from './external-link'
-import { TranslatedExternalLink } from './translated-external-link'
+import { ExternalLink } from '../../../links/external-link'
+import { TranslatedExternalLink } from '../../../links/translated-external-link'
 import { VersionInfo } from '../version-info/version-info'
 
 export const PoweredByLinks: React.FC = () => {
@@ -25,10 +25,7 @@ export const PoweredByLinks: React.FC = () => {
         Object.entries({ ...defaultLinks, ...(config.specialLinks) }).map(([i18nKey, href]) =>
           <Fragment key={i18nKey}>
             &nbsp;|&nbsp;
-            <TranslatedExternalLink
-              href={href}
-              i18nKey={i18nKey}
-            />
+            <TranslatedExternalLink href={href} i18nKey={i18nKey}/>
           </Fragment>
         )
       }
