+++ src/components/landing/layout/footer/social-links.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Trans, useTranslation } from 'react-i18next'
-import { ExternalLink } from './external-link'
+import { ExternalLink } from '../../../links/external-link'
 
 const SocialLink: React.FC = () => {
   useTranslation()
