+++ src/components/links/external-link.tsx
@@ -1,19 +1,20 @@
 import React, { Fragment } from 'react'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
-import { IconProp } from '../../../../utils/iconProp'
+import { IconProp } from '../../utils/iconProp'
 
 export interface ExternalLinkProp {
   href: string;
   text: string;
   icon?: IconProp;
+  className?: string
 }
 
-export const ExternalLink: React.FC<ExternalLinkProp> = ({ href, text, icon }) => {
+export const ExternalLink: React.FC<ExternalLinkProp> = ({ href, text, icon, className = 'text-light' }) => {
   return (
     <a href={href}
       target="_blank"
       rel="noopener noreferrer"
-      className="text-light">
+      className={className}>
       {
         icon
           ? <Fragment>
