+++ src/components/links/translated-external-link.tsx
@@ -1,16 +1,19 @@
 import React from 'react'
 import { useTranslation } from 'react-i18next'
+import { IconProp } from '../../utils/iconProp'
 import { ExternalLink } from './external-link'
 
 export interface TranslatedLinkProps {
-  href: string;
   i18nKey: string;
+  href: string;
+  icon?: IconProp;
+  className?: string
 }
 
-const TranslatedExternalLink: React.FC<TranslatedLinkProps> = ({ href, i18nKey }) => {
+const TranslatedExternalLink: React.FC<TranslatedLinkProps> = ({ i18nKey, ...props }) => {
   const { t } = useTranslation()
   return (
-    <ExternalLink href={href} text={t(i18nKey)}/>
+    <ExternalLink text={t(i18nKey)} {...props}/>
   )
 }
